training/two_phase/requirements_two_phase.txt
azure-storage-blob
cupy-cuda114
h5py
matplotlib
python-dotenv
zarr
